+++ app/javascript/components/charts/cp-performance/CPPerformance.js
@@ -9,30 +9,13 @@ import HighchartsReact from 'highcharts-react-official';
 
 import PlusIcon from 'images/icons/plus.svg';
 
+import { useOutsideClick } from 'shared/hooks';
 import { groupAllAreaSeries } from './helpers';
 import CompanySelector from './CompanySelector';
 import CompanyTag from './CompanyTag';
 import Tooltip from './Tooltip';
 
 const COLORS = ['#00C170', '#ED3D4A', '#FFDD49', '#440388', '#FF9600', '#B75038', '#00A8FF', '#F78FB3', '#191919', '#F602B4'];
-
-// TODO: move to common hooks
-const useCallbackOutsideClick = (element, action) => {
-  if (typeof action !== 'function') throw new Error('useCallbackOutsideClick expects action to be function');
-
-  const handleClickOutside = event => {
-    if (!element.current) return;
-    if (!element.current.contains(event.target)) action();
-  };
-
-  useEffect(() => {
-    document.addEventListener('mousedown', handleClickOutside);
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, []);
-};
 
 function getLegendItems(data) {
   return applyColorsToLegendItems(
@@ -62,7 +45,7 @@ function CPPerformance({ dataUrl, companySelector, unit }) {
         setLegendItems(getLegendItems(chartData));
       });
   }, []);
-  useCallbackOutsideClick(companySelectorWrapper, () => setCompanySelectorVisible(false));
+  useOutsideClick(companySelectorWrapper, () => setCompanySelectorVisible(false));
 
   const handleLegendItemRemove = (item) => {
     setLegendItems(
