+++ app/javascript/shared/hooks/index.js
@@ -1,0 +1,2 @@
+export { default as useDebounce } from './use-debounce';
+export { default as useOutsideClick } from './use-outside-click';
